[B2BORG-91]
feat: added the tab navigation on the organization's admin page
