+++ react/package.json
@@ -11,6 +11,8 @@
     "@types/jest": "^25.1.4",
     "@types/node": "^13.9.8",
     "@types/react": "^16.9.31",
+    "@types/react-router": "^5.1.11",
+    "@types/react-router-dom": "^5.1.11",
     "@vtex/test-tools": "^3.3.2",
     "@vtex/tsconfig": "^0.6.0",
     "apollo-cache-inmemory": "^1.6.5",
@@ -34,6 +36,8 @@
     "apollo-client": "^2.6.10",
     "react": "^16.9.2",
     "react-apollo": "^3.1.5",
-    "react-intl": "^5.13.4"
+    "react-intl": "^5.13.4",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2"
   }
 }
